Update README.md
Hexart.In AI Center was established in 2020 as part of the "Year of AI" initiative by the Telangana State Government. It is part of a CSR initiative by Hexagon Capability Center, India under an MoU with the Telangana State Government, NASSCOM Foundation, and Li2 Technologies.

Hexagon, in order to continue its “Education” outreach, decided to launch the AI Community Center with the aim to bring the do-it-yourself culture to the forefront, encouraging and motivating the school students and youth to become solution providers. The pedagogy is designed to engage them in self-learning, innovations, and inventions. In addition, as is the case associated with such skill development, one could expect the youth to develop higher-order thinking skills along with future world skills. critical vision to provide access to technology,  AI education, and opportunity to innovate for the community.

Currently, in its 3rd year of successful operations, Hexart has an outreach of 3847+ students; 1783+ Certifications, and 23+ active innovation project programs.